Cache pallets instead of calculating on-demand
