+++ Content.Client/_ES/Cargo/Storeroom/ESStoreroomSystem.cs
@@ -1,0 +1,5 @@
+using Content.Shared._ES.Cargo.Storeroom;
+
+namespace Content.Client._ES.Cargo.Storeroom;
+
+public sealed class ESStoreroomSystem : ESSharedStoreroomSystem;
