+++ Resources/Locale/en-US/_ES/commands/toolshed-descriptions.ftl
@@ -4,3 +4,7 @@ command-description-cast-view =
     Views the relationship data of this castmember.
 command-description-cast-viewAll =
     Views the relationship data of all castmembers.
+command-description-esstoreroom-viewStock =
+    Displays the warehouse stock on the station, using the cached value.
+command-description-esstoreroom-calculateStock =
+    Displays the warehouse stock on the station, calculating it on the fly.
