+++ Resources/Prototypes/Entities/Objects/Specific/Cargo/cargo_pallet.yml
@@ -5,11 +5,6 @@
   description: Common fixture of logistics and cargo. Subtle reminder where crates go during transport to avoid bruised shins.
   abstract: true
   components:
-  - type: ESStoreroomPallet #TODO: testing
-    goodsWhitelist:
-      components:
-      - GasCanister
-      - EntityStorage
   - type: InteractionOutline
   - type: Anchorable
     flags:
