+++ Resources/Prototypes/_ES/Entities/Structures/Specific/Cargo/pallet.yml
@@ -1,0 +1,44 @@
+- type: entity
+  parent: BaseStructure
+  id: ESBaseWarehousePallet
+  name: warehouse pallet
+  description: A highly advanced construct of plywood, capable of
+  components:
+  - type: Sprite
+    drawdepth: FloorTiles
+    layers:
+    - sprite: _ES/Structures/Specific/Cargo/pallet.rsi
+      state: icon
+  - type: ESStoreroomPallet
+    goodsWhitelist:
+      components:
+      - GasCanister
+      - EntityStorage
+  - type: InteractionOutline
+  - type: Anchorable
+  - type: CollideOnAnchor
+    enable: true
+  - type: Physics
+    canCollide: false
+  - type: Fixtures
+    fixtures:
+      fix1:
+        shape:
+          !type:PhysShapeAabb
+          bounds: "-0.45,-0.45,0.45,0.45"
+        density: 15
+        hard: false
+        mask:
+        - Impassable
+        - HighImpassable
+        - LowImpassable
+        layer:
+        - Impassable
+        - HighImpassable
+        - LowImpassable
+  - type: Damageable
+    damageContainer: StructuralInorganic
+    damageModifierSet: Wood
+  - type: GuideHelp
+    guides:
+    - Cargo
